tech?(api): move user to certif on correct model
