+++ api/tests/identity-access-management/unit/domain/models/User.test.js
@@ -654,39 +654,4 @@ describe('Unit | Identity Access Management | Domain | Model | User', function (
       expect(anonymizedUser.updatedAt.toISOString()).to.equal('2023-03-01T00:00:00.000Z');
     });
   });
-
-  describe('#has', function () {
-    it('should return false when user has no organization learner ids at all', function () {
-      // given
-      const user = domainBuilder.certification.enrolment.buildUser({ organizationLearnerIds: [] });
-
-      // when
-      const hasOrgaLearnerId = user.has({ organizationLearnerId: 123 });
-
-      // then
-      expect(hasOrgaLearnerId).to.be.false;
-    });
-
-    it('should return false when user does not have the given organizationLearnerId in its ids', function () {
-      // given
-      const user = domainBuilder.certification.enrolment.buildUser({ organizationLearnerIds: [456, 789] });
-
-      // when
-      const hasOrgaLearnerId = user.has({ organizationLearnerId: 123 });
-
-      // then
-      expect(hasOrgaLearnerId).to.be.false;
-    });
-
-    it('should return true when user has the given organizationLearnerId in its ids', function () {
-      // given
-      const user = domainBuilder.certification.enrolment.buildUser({ organizationLearnerIds: [456, 789, 123] });
-
-      // when
-      const hasOrgaLearnerId = user.has({ organizationLearnerId: 123 });
-
-      // then
-      expect(hasOrgaLearnerId).to.be.true;
-    });
-  });
 });
